% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_s0.R
\name{calculate.s0}
\alias{calculate.s0}
\title{Return to this script after finalizing the simulate_population.R script. Want to emulate simualte.pop exactly, but calculate s0, so I'll just manually figure out s0 values that work until I get a reasonable version of the script running.}
\usage{
calculate.s0(
  input_data,
  num_mates,
  num_years,
  movement_array = NULL,
  popstructure = "panmictic",
  stickiness = NULL,
  sticky_age = NULL,
  sticky_interval = NULL,
  superpod_size = NULL,
  male_behavior = NULL,
  sampling = "random",
  sample_size = NULL,
  sample_years = NULL,
  process_by = "age"
)
}
\description{
Return to this script after finalizing the simulate_population.R script. Want to emulate simualte.pop exactly, but calculate s0, so I'll just manually figure out s0 values that work until I get a reasonable version of the script running.
}
