% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{create.pop.data}
\alias{create.pop.data}
\title{Define parameters for population simulation}
\usage{
create.pop.data(
  max_age,
  survival,
  pop_number,
  pop_size,
  mating_periodicity,
  maturity_age = NULL,
  litter_size = NULL,
  YOY_survival = NULL,
  stable = TRUE,
  F_by_age = NULL,
  female_fraction = 0.5,
  infertility = 0,
  growth_params = NULL
)
}
\arguments{
\item{max_age}{Integer. Maximum age class (inclusive), so ages are 0:max_age.}

\item{survival}{Numeric vector. If stable = TRUE, length must be max_age
(s1..s_max_age), and s0 is calculated. If stable = FALSE, length must
be max_age + 1, including s0.}

\item{pop_number}{Integer. Number of populations.}

\item{pop_size}{Integer vector (length = pop_number). Total size per population.}

\item{mating_periodicity}{Integer. 1 = annual, 2 = biennial, etc.}

\item{maturity_age}{Optional integer. Used to derive fecundity if F_by_age is NULL. We do not need maturity age of males for this.}

\item{litter_size}{Optional numeric > 0. Mean pups per litter.}

\item{YOY_survival}{Optional numeric. Overrides s0 if stable = FALSE.}

\item{stable}{Logical. If TRUE, solve for s0 using Euler-Lotka.}

\item{F_by_age}{Optional numeric vector (length max_age + 1). Direct fecundity.}

\item{female_fraction}{Numeric in (0,1]. Default 0.5.}

\item{growth_params}{Either a single dataframe or a list of length pop_number with each list element corresponding to a dataframe with growth parameters for the population. If a single dataframe is used, then the growth parameters will be applied to each population; if a list of length pop_number is defined, then population-specific growth parameters can be defined. Each dataframe must have two rows, which define the sex-specific vonBertalanffy growth parameters for the species/population. If defined, these parameters will be used to assign a growth rate to each individual in the simulation at birth based on random draws from a multivariate normal distribution. Thereafter, growth proceeds deterministically, governed by the assigned growth rate. This dataframe, which will be formatted and saved for the simulation, must have one row for males and one for females and contain the following named columns:
\itemize{
\item L_inf: point estimate of L_inf (asymptotic maximum length)
\item L_inf_sd: standard deviation of L_inf
\item K: point estimate of K (growth coefficient)
\item K_sd: standard deviation surrounding K
\item t0: theoretical age at zero length
\item rho: correlation coefficient between asymptotic length (L_inf) and growth rate (K), bounded between -1 and 1. A negative value for rho implies that individuals that grow fast (large K) tend to have smaller asymptotic maximum lengths (small L_inf), while individuals that grow slower (small K) have larger asymptotic maximum lengths (large L_inf). A positive value for rho implies that individuals that grow fast (large K) also tend to reach larger maximum lengths (large L_inf). A value of 0 implies no correlation.
\item min_L0: minimum length-at-birth, in the same units as L_inf
\item max_L0: maximum length-at-birth, in the same units as L_inf
\item sex: to which sex do the parameters in this row apply? ("M" or "F")
}}
}
\value{
A list containing:
\itemize{
\item numbers_at_age: dataframe
\item survival: full survival vector including age 0 and max age (which is set to 0)
\item fecundity: vector of fecundity-by-age
\item s0: age 0 survival
}

Currently, create.pop.data takes a single vector of survival and a single vector of fecundity, which are applied to all populations. Future updates will allow for population-specific survival and fecundity.
}
\description{
Let's forget about this function for now (June 2026) and assume that all info is input to calculate_s0 and simulate_population
}
\details{
Format and prepare input data for individual-based population simulation from user-defined input values.
}
\examples{
input_pop <- create.pop.data(max_age = 10, survival = c(rep(0.8, times = 10)), pop_number = 3, pop_size = c(1000, 20000, 50000), mating_periodicity = 2, maturity_age = 5, litter_size = 7)
input_pop
}
