% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{create.pop.data}
\alias{create.pop.data}
\title{Define parameters for population simulation}
\usage{
create.pop.data(
  max_age,
  survival,
  pop_number,
  pop_size,
  mating_periodicity,
  maturity_age = NULL,
  litter_size = NULL,
  YOY_survival = NULL,
  stable = TRUE,
  F_by_age = NULL,
  female_fraction = 0.5
)
}
\arguments{
\item{max_age}{Integer. Maximum age class (inclusive), so ages are 0:max_age.}

\item{survival}{Numeric vector. If stable = TRUE, length must be max_age
(s1..s_max_age), and s0 is calculated. If stable = FALSE, length must
be max_age + 1, including s0.}

\item{pop_number}{Integer. Number of populations.}

\item{pop_size}{Integer vector (length = pop_number). Total size per population.}

\item{mating_periodicity}{Integer. 1 = annual, 2 = biennial, etc.}

\item{maturity_age}{Optional integer. Used to derive fecundity if F_by_age is NULL.}

\item{litter_size}{Optional numeric > 0. Mean pups per litter.}

\item{YOY_survival}{Optional numeric. Overrides s0 if stable = FALSE.}

\item{stable}{Logical. If TRUE, solve for s0 using Euler–Lotka.}

\item{F_by_age}{Optional numeric vector (length max_age + 1). Direct fecundity.}

\item{female_fraction}{Numeric in (0,1]. Default 0.5.}
}
\value{
A list containing:
\itemize{
\item numbers_at_age: dataframe
\item survival: full survival vector including age 0 and setting survival of max age to 0
\item fecundity: fecundity vector used
\item s0: age 0 survival
}
}
\description{
Define parameters for population simulation
}
\examples{
create.pop.data(max_age = 10, survival = c(rep(0.8, times = max_age)), pop_number = 3, pop_size = c(1000, 20000, 50000), mating_periodicity = 2, maturity_age = 5, litter_size = 7)
}
